django/requirements.txt
# ===================== Django 核心（固定版本） =====================
Django>=5.2,<5.3

# ===================== 配置管理 =====================
django-environ

# ===================== 日志增强 =====================
colorlog
python-json-logger

# ===================== 数据库驱动（✅ 取消注释这个） =====================
psycopg[binary]  # 🔑 Django 5.2 官方推荐的新驱动

# ===================== 缓存/队列 =====================
django-redis
celery[redis]

# ===================== Windows 专用 =====================
eventlet; platform_system == "Windows"